task/cleanup-url-utils * remove calculateFetchAll, move calculateLogView into dashboard where it belongs, stop exporting applyFetchAll
